refacto (cargo): Rename io crate to arch.
Since this crate shall contains arch specific bits of code it makes more
sense to call it arch like almost everywhere.
