+++ Cargo.toml
@@ -7,7 +7,7 @@ edition = "2021"
 
 members = [
     "serial",
-    "io",
+    "arch",
 ]
 
 [dependencies]
@@ -15,7 +15,7 @@ bootloader = "0.9.8"
 volatile = "0.2.6"
 log = "0.4.14"
 serial = { path = "./serial", optional = true }
-io = { path = "./io", optional = true }
+arch = { path = "./arch", optional = true }
 
 
 [features]
