+++ arch/Cargo.toml
@@ -1,5 +1,5 @@
 [package]
-name = "io"
+name = "arch"
 version = "0.1.0"
 edition = "2021"
