+++ serial/src/registers/lsr.rs
@@ -1,5 +1,5 @@
 use super::{ReadRegister, Register};
-use crate::io::inb;
+use crate::arch::inb;
 use crate::ComPort;
 
 const LSR_OFFSET: u16 = 5;
