+++ serial/src/registers/sr.rs
@@ -1,5 +1,5 @@
 use super::{ReadRegister, Register, WriteRegister};
-use crate::io::{inb, outb};
+use crate::arch::{inb, outb};
 use crate::ComPort;
 
 const SR_OFFSET: u16 = 7;
